requirements.txt
paho-mqtt>=2.1.0
protobuf>=5.29.0
grpcio-tools>=1.68.0
amqtt>=0.11.2
